requirements.txt
fastapi==0.68.1
streamlit==1.4.0
requests==2.26.0
pandas==1.3.4
numpy==1.21.3
scikit-learn==1.0
python-dotenv==0.19.1
